% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{quinolonehira6-package}
\alias{quinolonehira6}
\alias{quinolonehira6-package}
\title{quinolonehira6: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
